+++ src/lib/spudnik.ts
@@ -96,7 +96,6 @@ export class Spudnik {
 				['random', 'Random'],
 				['ref', 'Reference'],
 				['roles', 'Role'],
-				['stats', 'Stats'],
 				['translate', 'Translate'],
 				['util', 'Utility'],
 				['util-required', 'Required Utility']
